PrivateRouter moved to a new file
